feat(breaking): renamed dataloader.py to loader.py created data module
