+++ .gitignore
@@ -1,5 +1,5 @@
 # segma specific files and folders
-data/
+/data/
 logs/
 wandb/
 notebooks/
