+++ src/segma/data/__init__.py
@@ -1,0 +1,4 @@
+from .file_dataset import SegmaFileDataset, URISubsetLeakageError
+from .loaders import SegmentationDataLoader
+
+__all__ = ["SegmentationDataLoader", "SegmaFileDataset", "URISubsetLeakageError"]
